% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bddtestcoverage-package.R
\docType{package}
\name{bddtestcoverage-package}
\alias{bddtestcoverage}
\alias{bddtestcoverage-package}
\title{bddtestcoverage: BDD test functions and coverage}
\description{
Package designed to explore BDD functions and test coverage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjfrigaard/bddtestcoverage}
}

}
\keyword{internal}
