% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_movies_genres.R
\name{collapse_movies_genres}
\alias{collapse_movies_genres}
\title{Test function (tidy ggplot2movies::movies)}
\usage{
collapse_movies_genres(movies_data)
}
\arguments{
\item{movies_data}{\code{ggplot2movies::movies} data}
}
\value{
A tidy \code{movies} \code{data.frame}
}
\description{
Test function (tidy ggplot2movies::movies)
}
\examples{
require(ggplot2movies)
collapse_movies_genres(ggplot2movies::movies)
}
