% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_movies_tidy.R
\name{tidy_ggp2movies}
\alias{tidy_ggp2movies}
\title{Test function (tidy ggplot2movies::movies)}
\usage{
tidy_ggp2movies(movies_data)
}
\arguments{
\item{movies_data}{\code{ggplot2movies::movies} data}
}
\value{
A tidy \code{movies} \code{data.frame}
}
\description{
Test function (tidy ggplot2movies::movies)
}
\examples{
require(ggplot2movies)
tidy_ggp2movies(ggplot2movies::movies)
}
