% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bddtestcoverage-package.R
\docType{package}
\name{bddtestcoverage-package}
\alias{bddtestcoverage}
\alias{bddtestcoverage-package}
\title{bddtestcoverage: Package to exercise tests under test coverage}
\description{
Designed to test BDD functions with test coverage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjfrigaard/bddtestcoverage}
  \item Report bugs at \url{https://github.com/mjfrigaard/bddtestcoverage/issues}
}

}
\keyword{internal}
